added provision for blank matched string to return a NONE
git-svn-id: http://usg.lofar.org/svn/code/trunk@5980 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/Pulsar/scripts/pulsars_at_location.sh
@@ -95,6 +95,13 @@ else
    matched_string=`cat $matched_file | sort -n -r -k 4,4 | head -$nmax | awk '{print $1}' |  tr '\n' ',' | sed 's/,$//g'`
 fi
 
+if [[ $matched_string == "" ]]
+then
+   echo "Position not matched to Pulsar Catalog"
+   matched_string="NONE"
+   exit
+fi
+
 echo $matched_string
 
 rm $matched_file
